[Rust]: Remove the unused functions.
