+++ bindings/rust/wasmedge-sys/Cargo.toml
@@ -21,7 +21,7 @@ rand = "0.8.4"
 thiserror = "1.0.30"
 wasmedge-macro = {path = "../wasmedge-macro", version = "0.1"}
 wasmedge-types = {path = "../wasmedge-types", version = "0.3"}
-wasmtime-fiber = "1"
+wasmtime-fiber = "2"
 
 [build-dependencies]
 bindgen = {version = "0.60.1", default-features = false, features = ["runtime"]}
@@ -29,6 +29,7 @@ cmake = "0.1"
 
 [dev-dependencies]
 tokio = {version = "1", features = ["full"]}
+anyhow = "1"
 
 [features]
 aot = []
