+++ bindings/rust/wasmedge-sys/examples/async_host_func.rs
@@ -1,3 +1,13 @@
+//! This example demonstrates how to create host functions and call them asynchronously.
+//!
+//! To run this example, use the following command:
+//!
+//! ```bash
+//! cd /wasmedge-root-dir/bindings/rust/
+//!
+//! cargo run -p wasmedge-sys --example async_host_func
+//! ```
+
 use wasmedge_sys::{Executor, FuncType, Function, Store};
 
 #[tokio::main]
@@ -9,7 +19,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // create a host function
     let async_host_func = Function::create_async(
         &func_ty,
-        Box::new(|_frame, _input, _data| {
+        |_frame, _input, _data| {
             Box::new(async {
                 println!("Hello, world!");
                 tokio::time::sleep(std::time::Duration::from_secs(1)).await;
@@ -25,7 +35,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 println!("Hello, world after sleep!");
                 Ok(vec![])
             })
-        }),
+        },
         Some(&mut store),
         0,
     )?;
