+++ bindings/rust/wasmedge-sys/examples/async_run_wasm_from_bytes.rs
@@ -1,3 +1,13 @@
+//! This example demonstrates how to load and run wasm from bytes asynchronously.
+//!
+//! To run this example, use the following command:
+//!
+//! ```bash
+//! cd /wasmedge-root-dir/bindings/rust/
+//!
+//! cargo run -p wasmedge-sys --example async_run_wasm_from_bytes
+//! ```
+
 use wasmedge_sys::{Config, Store, Vm, WasmValue};
 use wasmedge_types::wat2wasm;
 
@@ -57,8 +67,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     assert!(result.is_ok());
     let wasm_bytes = result.unwrap();
 
-    let fut1 = vm.run_wasm_from_bytes_async2(&wasm_bytes, "fib", [WasmValue::from_i32(20)]);
-    let fut2 = vm.run_wasm_from_bytes_async2(&wasm_bytes, "fib", [WasmValue::from_i32(5)]);
+    let fut1 = vm.run_wasm_from_bytes_async(&wasm_bytes, "fib", [WasmValue::from_i32(20)]);
+    let fut2 = vm.run_wasm_from_bytes_async(&wasm_bytes, "fib", [WasmValue::from_i32(5)]);
 
     let (ret1, ret2) = tokio::join!(fut1, fut2);
     let returns1 = ret1?;
